+++ lib/haml/filters.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: false
 require "tilt"
 
 module Haml
@@ -183,8 +184,7 @@ RUBY
 
           rendered = Haml::Helpers::find_and_preserve(filter.render_with_options(text, compiler.options), compiler.options[:preserve])
           rendered.rstrip!
-          rendered.gsub!("\n", "\n#{'  ' * @output_tabs}") unless options[:ugly]
-          push_text(rendered)
+          push_text("#{rendered}\n")
         end
       end
     end