+++ lib/haml/template.rb
@@ -1,5 +1,5 @@
+# frozen_string_literal: true
 require 'haml/template/options'
-require 'haml/engine'
 if defined?(ActiveSupport)
   ActiveSupport.on_load(:action_view) do
     require 'haml/helpers/action_view_mods'
@@ -13,7 +13,7 @@ require 'haml/helpers/xss_mods'
 require 'haml/helpers/action_view_xss_mods'
 
 module Haml
-  class Compiler
+  class TempleEngine
     def precompiled_method_return_value_with_haml_xss
       "::Haml::Util.html_safe(#{precompiled_method_return_value_without_haml_xss})"
     end
@@ -33,7 +33,6 @@ module Haml
 end
 
 
-Haml::Template.options[:ugly] = defined?(Rails) ? !Rails.env.development? : true
 Haml::Template.options[:escape_html] = true
 
-require 'haml/template/plugin'
+require 'haml/plugin'